Change function print_conv prototype

+++ holberton.h
@@ -28,7 +28,7 @@ int print_o(va_list args);
 int print_x(va_list args);
 int print_X(va_list args);
 int print_S(va_list args);
-char *print_conv (unsigned int num, int base, int cap);
+char *print_conv (unsigned long int num, int base, int cap);
 int print_r(va_list args);
 int print_p(va_list args);
 int print_R(va_list args);
